+++ packages/stream_chat_flutter/lib/src/message_widget/bottom_row.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:stream_chat_flutter/src/message_widget/sending_indicator_wrapper.dart';
+import 'package:stream_chat_flutter/src/message_widget/sending_indicator_builder.dart';
 import 'package:stream_chat_flutter/src/message_widget/thread_painter.dart';
 import 'package:stream_chat_flutter/src/message_widget/thread_participants.dart';
 import 'package:stream_chat_flutter/src/message_widget/username.dart';
@@ -204,7 +204,7 @@ class BottomRow extends StatelessWidget {
       if (showSendingIndicator)
         WidgetSpan(
           child: sendingIndicatorBuilder?.call(context, message) ??
-              SendingIndicatorWrapper(
+              SendingIndicatorBuilder(
                 messageTheme: messageTheme,
                 message: message,
                 hasNonUrlAttachments: hasNonUrlAttachments,