+++ packages/stream_chat_flutter/lib/src/message_widget/sending_indicator_builder.dart
@@ -6,15 +6,16 @@ import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 ///
 /// Used in [BottomRow]. Should not be used elsewhere.
 /// {@endtemplate}
-class SendingIndicatorWrapper extends StatelessWidget {
+class SendingIndicatorBuilder extends StatelessWidget {
   /// {@macro sendingIndicatorWrapper}
-  const SendingIndicatorWrapper({
+  const SendingIndicatorBuilder({
     super.key,
     required this.messageTheme,
     required this.message,
     required this.hasNonUrlAttachments,
     required this.streamChat,
     required this.streamChatTheme,
+    this.channel,
   });
 
   /// {@macro messageTheme}
@@ -32,10 +33,14 @@ class SendingIndicatorWrapper extends StatelessWidget {
   /// {@macro streamChatThemeData}
   final StreamChatThemeData streamChatTheme;
 
+  /// {@macro channel}
+  final Channel? channel;
+
   @override
   Widget build(BuildContext context) {
     final style = messageTheme.createdAtStyle;
-    final memberCount = StreamChannel.of(context).channel.memberCount ?? 0;
+    final channel = this.channel ?? StreamChannel.of(context).channel;
+    final memberCount = channel.memberCount ?? 0;
 
     if (hasNonUrlAttachments &&
         (message.status == MessageSendingStatus.sending ||
@@ -58,8 +63,6 @@ class SendingIndicatorWrapper extends StatelessWidget {
       );
     }
 
-    final channel = StreamChannel.of(context).channel;
-
     return BetterStreamBuilder<List<Read>>(
       stream: channel.state?.readStream,
       initialData: channel.state?.read,
@@ -68,12 +71,14 @@ class SendingIndicatorWrapper extends StatelessWidget {
             it.user.id != streamChat.currentUser?.id &&
             (it.lastRead.isAfter(message.createdAt) ||
                 it.lastRead.isAtSameMomentAs(message.createdAt)));
-        final isMessageRead = readList.length >= (channel.memberCount ?? 0) - 1;
+
+        final isMessageRead = readList.isNotEmpty;
         Widget child = StreamSendingIndicator(
           message: message,
           isMessageRead: isMessageRead,
           size: style!.fontSize,
         );
+
         if (isMessageRead) {
           child = Row(
             mainAxisSize: MainAxisSize.min,
@@ -90,6 +95,7 @@ class SendingIndicatorWrapper extends StatelessWidget {
             ],
           );
         }
+
         return child;
       },
     );